+++ src/client/profile-provider.ts
@@ -1,4 +1,11 @@
-import { MapDocumentNode, ProfileDocumentNode } from '@superfaceai/ast';
+import {
+  assertMapDocumentNode,
+  assertProfileDocumentNode,
+  isMapFile,
+  isProfileFile,
+  MapDocumentNode,
+  ProfileDocumentNode,
+} from '@superfaceai/ast';
 import createDebug from 'debug';
 import { promises as fsp } from 'fs';
 import { join as joinPath } from 'path';
@@ -12,6 +19,7 @@ import {
 import {
   invalidProfileError,
   invalidSecurityValuesError,
+  referencedFileNotFoundError,
   securityNotFoundError,
   serviceNotFoundError,
 } from '../internal/errors.helpers';
@@ -28,6 +36,7 @@ import {
   mergeVariables,
   NonPrimitive,
 } from '../internal/interpreter/variables';
+import { Parser } from '../internal/parser';
 import {
   FILE_URI_PROTOCOL,
   isApiKeySecurityValues,
@@ -59,6 +68,7 @@ function profileAstId(ast: ProfileDocumentNode): string {
 const boundProfileProviderDebug = createDebug(
   'superface:bound-profile-provider'
 );
+
 export class BoundProfileProvider {
   //TODO: Interceptable and set metadata
   private profileValidator: ProfileParameterValidator;
@@ -177,7 +187,12 @@ export type BindConfiguration = {
 };
 
 const profileProviderDebug = createDebug('superface:profile-provider');
+
 export class ProfileProvider {
+  private profileId: string;
+  private scope: string | undefined;
+  private profileName: string;
+
   constructor(
     /** Preloaded superJson instance */
     //TODO: Use superJson from events/Client?
@@ -189,7 +204,22 @@ export class ProfileProvider {
     private events: Events,
     /** url or ast node */
     private map?: string | MapDocumentNode
-  ) {}
+  ) {
+    if (this.profile instanceof ProfileConfiguration) {
+      this.profileId = this.profile.id;
+    } else if (typeof this.profile === 'string') {
+      this.profileId = this.profile;
+    } else {
+      this.profileId = profileAstId(this.profile);
+    }
+    const [scopeOrProfileName, profileName] = this.profileId.split('/');
+    if (profileName === undefined) {
+      this.profileName = scopeOrProfileName;
+    } else {
+      this.scope = scopeOrProfileName;
+      this.profileName = profileName;
+    }
+  }
 
   /**
    * Binds the provider.
@@ -202,16 +232,7 @@ export class ProfileProvider {
     // resolve profile locally
     const profileAst = await this.resolveProfileAst();
     if (profileAst === undefined) {
-      let profileId;
-      if (this.profile instanceof ProfileConfiguration) {
-        profileId = this.profile.id;
-      } else if (typeof this.profile === 'string') {
-        profileId = this.profile;
-      } else {
-        profileId = profileAstId(this.profile);
-      }
-
-      throw invalidProfileError(profileId);
+      throw invalidProfileError(this.profileId);
     }
     const profileId = profileAstId(profileAst);
 
@@ -293,7 +314,18 @@ export class ProfileProvider {
 
     const profileAst = await ProfileProvider.resolveValue(
       resolveInput,
-      fileContents => JSON.parse(fileContents) as ProfileDocumentNode, // TODO: validate
+      async (fileContents, fileName) => {
+        // If we have profile source, we parse
+        if (fileName !== undefined && isProfileFile(fileName)) {
+          return Parser.parseProfile(fileContents, fileName, {
+            profileName: this.profileName,
+            scope: this.scope,
+          });
+        }
+
+        // Otherwise we return parsed
+        return assertProfileDocumentNode(JSON.parse(fileContents));
+      },
       profileId => {
         const profileSettings = this.superJson.normalized.profiles[profileId];
         if (profileSettings === undefined) {
@@ -302,23 +334,19 @@ export class ProfileProvider {
         } else if ('file' in profileSettings) {
           // assumed right next to source file
           return (
-            FILE_URI_PROTOCOL +
-            this.superJson.resolvePath(profileSettings.file) +
-            '.ast.json'
+            FILE_URI_PROTOCOL + this.superJson.resolvePath(profileSettings.file)
           );
         } else {
           // assumed to be in grid folder
           return (
             FILE_URI_PROTOCOL +
             this.superJson.resolvePath(
-              joinPath(
-                'grid',
-                profileId + `@${profileSettings.version}.supr.ast.json`
-              )
+              joinPath('grid', `${profileId}@${profileSettings.version}.supr`)
             )
           );
         }
-      }
+      },
+      ['.ast.json', '']
     );
 
     return profileAst;
@@ -335,7 +363,7 @@ export class ProfileProvider {
 
     const providerInfo = await ProfileProvider.resolveValue<ProviderJson>(
       resolveInput,
-      fileContents => JSON.parse(fileContents) as ProviderJson, // TODO: validate
+      async fileContents => JSON.parse(fileContents) as ProviderJson, // TODO: validate
       providerName => {
         const providerSettings =
           this.superJson.normalized.providers[providerName];
@@ -371,9 +399,22 @@ export class ProfileProvider {
     mapRevision?: string;
   }> {
     const mapInfo: { mapVariant?: string; mapRevision?: string } = {};
+    const [, providerName] = mapId.split('.');
     const mapAst = await ProfileProvider.resolveValue<MapDocumentNode>(
       this.map ?? mapId,
-      fileContents => JSON.parse(fileContents) as MapDocumentNode, // TODO: validate
+      async (fileContents, fileName) => {
+        // If we have source, we parse
+        if (fileName !== undefined && isMapFile(fileName)) {
+          return Parser.parseMap(fileContents, fileName, {
+            profileName: this.profileName,
+            providerName,
+            scope: this.scope,
+          });
+        }
+
+        // Otherwise we return parsed
+        return assertMapDocumentNode(JSON.parse(fileContents));
+      },
       mapId => {
         const [profileId, providerName] = mapId.split('.');
         const profileProviderSettings =
@@ -384,8 +425,7 @@ export class ProfileProvider {
         } else if ('file' in profileProviderSettings) {
           return (
             FILE_URI_PROTOCOL +
-            this.superJson.resolvePath(profileProviderSettings.file) +
-            '.ast.json'
+            this.superJson.resolvePath(profileProviderSettings.file)
           );
         } else {
           mapInfo.mapVariant = profileProviderSettings.mapVariant;
@@ -393,7 +433,8 @@ export class ProfileProvider {
 
           return undefined;
         }
-      }
+      },
+      ['.ast.json', '']
     );
 
     return {
@@ -413,24 +454,36 @@ export class ProfileProvider {
    */
   private static async resolveValue<T>(
     input: T | string | undefined,
-    parseFile: (contents: string) => T,
-    unpackNested: (input: string) => T | string | undefined
+    parseFile: (contents: string, fileName?: string) => Promise<T>,
+    unpackNested: (input: string) => T | string | undefined,
+    extensions: string[] = ['']
   ): Promise<T | undefined> {
     if (typeof input === 'string') {
       if (isFileURIString(input)) {
-        profileProviderDebug('Resolving input as file:', input);
+        const fileName = input.slice(FILE_URI_PROTOCOL.length);
+        profileProviderDebug('Resolving input as file:', fileName);
 
         // read in files
-        return parseFile(
-          await fsp.readFile(input.slice(FILE_URI_PROTOCOL.length), {
-            encoding: 'utf-8',
-          })
-        );
-        // eslint-disable-next-line no-constant-condition
-      } else if (false) {
-        profileProviderDebug('Resolving input as url:', input);
-        // TODO: detect remote url and fetch it, or call a callback?
+        let contents, fileNameWithExtension;
+        for (const extension of extensions) {
+          fileNameWithExtension = fileName + extension;
+          try {
+            contents = await fsp.readFile(fileNameWithExtension, {
+              encoding: 'utf-8',
+            });
+            break;
+          } catch (e) {
+            void e;
+          }
+        }
+
+        if (contents === undefined) {
+          throw referencedFileNotFoundError(fileName, extensions);
+        }
+
+        return parseFile(contents, fileNameWithExtension);
       } else {
+        // TODO: detect remote url and fetch it, or call a callback?
         profileProviderDebug('Resolving input as nested value:', input);
         // unpack nested and recursively process them
         const nested = unpackNested(input);